Scripts
These are the scripts for the game. They may have been included in the previous update but these are the newest ones. You will need to set up the dependencies such as adding ShipControl to the player ship. If something does not work message me and I will walk you through setting everything up